// src/models/IoK8sApiFlowcontrolV1beta3LimitResponse/types.ts
import { IoK8sApiFlowcontrolV1beta3QueuingConfigurationModelOptions as IoK8sApiFlowcontrolV1beta3QueuingConfiguration } from '../IoK8sApiFlowcontrolV1beta3QueuingConfiguration';

export interface IoK8sApiFlowcontrolV1beta3LimitResponseModelOptions {
  queuing?: IoK8sApiFlowcontrolV1beta3QueuingConfiguration;
  type: string;
}